+++ handler/file.go
@@ -37,7 +37,9 @@ type FileHandlerOptions struct {
 	FileMode fs.FileMode
 
 	// Level is the minimum log level to write to the handler.
-	Level slogx.Level
+	//
+	// If this is nil, it defaults to slogx.LevelInfo.
+	Level *slogx.LevelVar
 
 	// MaxFileCount indicates the maximum number of log files to keep, including the active log file.
 	//
@@ -62,7 +64,7 @@ func DefaultFileHandlerOptions() FileHandlerOptions {
 	return FileHandlerOptions{
 		DirMode:         0755,
 		FileMode:        0640,
-		Level:           slogx.LevelInfo,
+		Level:           slogx.NewLevelVar(slogx.LevelInfo),
 		MaxFileCount:    5,
 		MaxFileSize:     10000000,
 		RecordFormatter: formatter.DefaultJSONFormatter(),
@@ -113,6 +115,9 @@ func NewFileHandler(opts FileHandlerOptions) (*fileHandler, error) {
 	if opts.FileMode == 0 {
 		opts.FileMode = 0640
 	}
+	if opts.Level == nil {
+		opts.Level = slogx.NewLevelVar(slogx.LevelInfo)
+	}
 	if opts.MaxFileCount == 0 {
 		opts.MaxFileCount = 5
 	}
@@ -133,7 +138,7 @@ func NewFileHandler(opts FileHandlerOptions) (*fileHandler, error) {
 
 // Enabled determines whether or not the given level is enabled in this handler.
 func (h fileHandler) Enabled(ctx context.Context, level slog.Level) bool {
-	return level >= h.options.Level.Level()
+	return slogx.Level(level) >= h.options.Level.Level()
 }
 
 // Handle actually handles writing the record to the file.
@@ -160,11 +165,6 @@ func (h *fileHandler) Handle(ctx context.Context, r slog.Record) error {
 
 	// write the buffer to the file
 	return h.write(buf)
-}
-
-// Level returns the current logging level that is in use by the handler.
-func (h fileHandler) Level() slogx.Level {
-	return h.options.Level
 }
 
 // Shutdown is responsible for cleaning up resources used by the handler.
@@ -209,21 +209,6 @@ func (h fileHandler) WithGroup(name string) slog.Handler {
 		newHandler.activeGroup = name
 	}
 	return newHandler
-}
-
-// WithLevel returns a new handler with the given logging level set.
-func (h fileHandler) WithLevel(level slogx.Level) slogx.DynamicLevelHandler {
-	options := h.options
-	options.Level = level
-	return &fileHandler{
-		activeGroup:     h.activeGroup,
-		attrs:           h.attrs,
-		currentFileSize: h.currentFileSize,
-		file:            h.file,
-		groups:          h.groups,
-		options:         options,
-		writeLock:       h.writeLock,
-	}
 }
 
 // openFile opens the log file for writing or creates it and any parent folders if they do not exist.