+++ handler/json.go
@@ -19,7 +19,9 @@ type jsonHandlerOptionsContext struct{}
 // JSONHandlerOptions holds the options for the JSON handler.
 type JSONHandlerOptions struct {
 	// Level is the minimum log level to write to the handler.
-	Level slogx.Level
+	//
+	// If this is nil, it defaults to slogx.LevelInfo.
+	Level *slogx.LevelVar
 
 	// RecordFormatter specifies the formatter to use to format the record before writing it to the writer.
 	//
@@ -35,7 +37,7 @@ type JSONHandlerOptions struct {
 // DefaultJSONHandlerOptions returns a default set of options for the handler.
 func DefaultJSONHandlerOptions() JSONHandlerOptions {
 	return JSONHandlerOptions{
-		Level:           slogx.LevelInfo,
+		Level:           slogx.NewLevelVar(slogx.LevelInfo),
 		RecordFormatter: formatter.DefaultJSONFormatter(),
 		Writer:          os.Stdout,
 	}
@@ -72,6 +74,9 @@ type jsonHandler struct {
 // NewJSONHandler creates a new handler object.
 func NewJSONHandler(opts JSONHandlerOptions) *jsonHandler {
 	// set default options
+	if opts.Level == nil {
+		opts.Level = slogx.NewLevelVar(slogx.LevelInfo)
+	}
 	if opts.Writer == nil {
 		opts.Writer = os.Stdout
 	}
@@ -87,7 +92,7 @@ func NewJSONHandler(opts JSONHandlerOptions) *jsonHandler {
 
 // Enabled determines whether or not the given level is enabled in this handler.
 func (h jsonHandler) Enabled(ctx context.Context, level slog.Level) bool {
-	return level >= h.options.Level.Level()
+	return slogx.Level(level) >= h.options.Level.Level()
 }
 
 // Handle actually handles writing the record to the output writer.
@@ -117,11 +122,6 @@ func (h *jsonHandler) Handle(ctx context.Context, r slog.Record) error {
 	defer h.writeLock.Unlock()
 	_, err = h.options.Writer.Write(buf.Bytes())
 	return err
-}
-
-// Level returns the current logging level that is in use by the handler.
-func (h jsonHandler) Level() slogx.Level {
-	return h.options.Level
 }
 
 // Shutdown is responsible for cleaning up resources used by the handler.
@@ -164,17 +164,4 @@ func (h *jsonHandler) WithGroup(name string) slog.Handler {
 		newHandler.activeGroup = name
 	}
 	return newHandler
-}
-
-// WithLevel returns a new handler with the given logging level set.
-func (h jsonHandler) WithLevel(level slogx.Level) slogx.DynamicLevelHandler {
-	options := h.options
-	options.Level = level
-	return &jsonHandler{
-		activeGroup: h.activeGroup,
-		attrs:       h.attrs,
-		groups:      h.groups,
-		options:     options,
-		writeLock:   h.writeLock,
-	}
 }