+++ handler/nil.go
@@ -3,8 +3,6 @@ package handler
 import (
 	"context"
 	"log/slog"
-
-	"go.innotegrity.dev/slogx"
 )
 
 // nilHandler simply discards all messages.
@@ -29,13 +27,6 @@ func (h *nilHandler) Handle(ctx context.Context, r slog.Record) error {
 	return nil
 }
 
-// Level resturns the current logging level that is in use by the handler.
-//
-// This function always returns slogx.LevelNone.
-func (h nilHandler) Level() slogx.Level {
-	return slogx.LevelNone
-}
-
 // WithAttrs creates a new handler from the existing one adding the given attributes to it.
 //
 // This function always just returns a new, default NilHandler object.
@@ -47,12 +38,5 @@ func (h nilHandler) WithAttrs(attrs []slog.Attr) slog.Handler {
 //
 // This function always just returns a new, default NilHandler object.
 func (h nilHandler) WithGroup(name string) slog.Handler {
-	return NewNilHandler()
-}
-
-// WithLevel creates a new handler from the existing one setting the level to the given level.
-//
-// This function always just returns a new, default NilHandler object.
-func (h nilHandler) WithLevel(l slogx.Level) slogx.DynamicLevelHandler {
 	return NewNilHandler()
 }